Contributes to #1297.

@krizhanovsky said very well:

len in ss_skb_actor_t() comes from skb_headlen(), which is unsigned int. unsigned int headlen also makes the code more robust.

What was done here:
 - used unsigned int instead of size_t in ss_skb_actor_t;
 - also unsigned int is used instead of size_t at major network data processing points.

+++ fw/t/bomber.c
@@ -159,7 +159,7 @@ tfw_bmb_conn_drop(struct sock *sk)
 }
 
 static int
-tfw_bmb_print_msg(void *msg_data, unsigned char *data, size_t len,
+tfw_bmb_print_msg(void *msg_data, unsigned char *data, unsigned int len,
 		  unsigned int *read)
 {
 	printk(KERN_INFO "%.*s\n", (int)len, data);
