+++ tls/ttls.c
@@ -2134,7 +2134,7 @@ ttls_handshake_step(TlsCtx *tls, unsigned char *buf, size_t len, size_t hh_len,
  * The function adds the number of bytes parsed in @buf to @read.
  */
 int
-ttls_recv(void *tls_data, unsigned char *buf, size_t len, unsigned int *read)
+ttls_recv(void *tls_data, unsigned char *buf, unsigned int len, unsigned int *read)
 {
 	int r;
 	unsigned int hh_len = 0, parsed = *read;
